refactor: mv cmds out of cmd/ and introduce bpaf opts.rs
