+++ src/auth.rs
