--- src/cmd/mod.rs
@@ -1,6 +1,0 @@
-pub mod auth;
-pub mod commit;
-pub mod find;
-pub mod rebase;
-pub mod reword;
-pub mod status;
