+++ src/providers/provider.rs
@@ -18,7 +18,6 @@ use super::{
     Display,
     Serialize,
     Deserialize,
-    clap::ValueEnum,
 )]
 pub enum ProviderKind {
     OpenAI,
